#6 Last Update

+++ src/main/java/com/kodhnk/base/dataAccess/ExaminationRepository.java
@@ -3,5 +3,5 @@ package com.kodhnk.base.dataAccess;
 import com.kodhnk.base.entities.Examination;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExaminationRepository extends JpaRepository<Examination, Integer> {
+public interface ExaminationRepository extends JpaRepository<Examination, Long> {
 }
