+++ src/main/java/com/kodhnk/base/dataAccess/HospitalRepository.java
@@ -3,5 +3,5 @@ package com.kodhnk.base.dataAccess;
 import com.kodhnk.base.entities.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
+public interface HospitalRepository extends JpaRepository<Hospital, Long> {
 }
